Relax requirements for payment_product
Validation in the app has been causing some problems, so this has been
relaxed a bit so that it can be quickly disabled without the gateway
completely breaking.

During testing the gateway seemed to periodically generate errors if the
product type and credit card numbers didn't agree, but it would
sometimes let cards through which mismatched. Odd.